create evidence modal
